+++ modules/core/src/main/java/io/github/palexdev/mfxcore/observables/OnInvalidated.java
@@ -192,8 +192,8 @@ public class OnInvalidated<T> extends When<T> {
 		}
 
 		invalidatingObservables.forEach(o -> o.addListener(invalidationListener));
+		register();
 		observableValue.addListener(listener);
-		whens.put(observableValue, this);
 		return this;
 	}